Update OpenCV to 4.18.X and allow numpy 2.X (#366)
### Motivation
We are still using numpy 1.X. This is because we were stuck at on opencv
4.8 since newer versions at the time had issues with the omnidir camera
calibration module. This seems to have been fixed and as such we want to
allow newer numpy versions and a new opencv minor version.

### Implementation

<!-- What is the concept behind the implementation? How does it work?
-->
<!-- Include any important technical decisions or trade-offs made -->
- The numpy dependency is now `<3.0`.
- The opencv-python-contrib dependency is now `>=4.12.0.88` (current
newest) and `<4.13.0.0`.
- opencv-python was removed (not needed with opencv-python-contrib)
- cv2 is ignored by mypy (installing stubs takes more than 5 minutes)
- Tests for real hardware cameras were fixed (likely unrelated, but was
not a big fix)

---------

+++ rosys/hardware/bms.py
@@ -130,7 +130,7 @@ class BmsHardware(Bms, ModuleHardware):
         self.state.percentage = percentage
         self.state.voltage = result.get('total voltage')
         self.state.current = result.get('current')
-        self.state.temperature = np.mean(result['temperatures']) if 'temperatures' in result else None
+        self.state.temperature = float(np.mean(result['temperatures'])) if 'temperatures' in result else None
         is_charging = (self.state.current or 0) > self.charge_detect_threshold
         self._send_charging_events(is_charging)
         self.state.is_charging = is_charging
